+++ CPU/CPUsoftware/src/tools/CpuTools.h
@@ -13,6 +13,7 @@
 
 #include <dirent.h>
 #include <unistd.h>
+#include <stdint.h>
 
 #include "ZynqManager.h"
 
@@ -35,7 +36,7 @@ public:
   static std::string SpaceToUnderscore(std::string);
   static bool PingConnect(std::string ip_address);
   static bool CheckFtp();
-  static std::vector<int> DelimStrToVec(std::string input_string, char delim, int size, bool check_01);
+  static std::vector<int> DelimStrToVec(std::string input_string, char delim, uint8_t size, bool check_01);
   static std::string BuildStr(std::string stem, std::string sep, int val, int rep);
   static std::string BuildStrFromVec(std::string stem, std::string sep, std::vector<int> values);
